Convert Round to Round Interface, leaving Room for an eventual tenth round implementation
